fix bugs in np.zeros for depth-wise conv

+++ feature_extractor.py
@@ -22,7 +22,7 @@ def feature_extractor(act0, params,
     # [2] conv
     weight2 = ng.variable(dtype=weight_dtype, shape=(32, 3, 3, 32), name="layer1.3.weight")
     weight2_value_org = params["layer1.3.weight"]
-    weight2_value = np.zeros((32, 3, 3, 32))
+    weight2_value = np.zeros((32, 3, 3, 32), dtype=np.int8)
     for i, j, k in np.ndindex((32, 3, 3)):
         weight2_value[i][j][k][i] = weight2_value_org[i][j][k][0]
     weight2.set_value(weight2_value)
@@ -68,7 +68,7 @@ def feature_extractor(act0, params,
     # [5] conv
     weight5 = ng.variable(dtype=weight_dtype, shape=(48, 3, 3, 48), name="layer2.0.0.layers.3.weight")
     weight5_value_org = params["layer2.0.0.layers.3.weight"]
-    weight5_value = np.zeros((48, 3, 3, 48))
+    weight5_value = np.zeros((48, 3, 3, 48), dtype=np.int8)
     for i, j, k in np.ndindex((48, 3, 3)):
         weight5_value[i][j][k][i] = weight5_value_org[i][j][k][0]
     weight5.set_value(weight5_value)
@@ -114,7 +114,7 @@ def feature_extractor(act0, params,
     # [8] conv
     weight8 = ng.variable(dtype=weight_dtype, shape=(72, 3, 3, 72), name="layer2.0.1.layers.3.weight")
     weight8_value_org = params["layer2.0.1.layers.3.weight"]
-    weight8_value = np.zeros((72, 3, 3, 72))
+    weight8_value = np.zeros((72, 3, 3, 72), dtype=np.int8)
     for i, j, k in np.ndindex((72, 3, 3)):
         weight8_value[i][j][k][i] = weight8_value_org[i][j][k][0]
     weight8.set_value(weight8_value)
@@ -164,7 +164,7 @@ def feature_extractor(act0, params,
     # [12] conv
     weight12 = ng.variable(dtype=weight_dtype, shape=(72, 3, 3, 72), name="layer2.0.2.layers.3.weight")
     weight12_value_org = params["layer2.0.2.layers.3.weight"]
-    weight12_value = np.zeros((72, 3, 3, 72))
+    weight12_value = np.zeros((72, 3, 3, 72), dtype=np.int8)
     for i, j, k in np.ndindex((72, 3, 3)):
         weight12_value[i][j][k][i] = weight12_value_org[i][j][k][0]
     weight12.set_value(weight12_value)
@@ -214,7 +214,7 @@ def feature_extractor(act0, params,
     # [16] conv
     weight16 = ng.variable(dtype=weight_dtype, shape=(72, 5, 5, 72), name="layer3.0.0.layers.3.weight")
     weight16_value_org = params["layer3.0.0.layers.3.weight"]
-    weight16_value = np.zeros((72, 5, 5, 72))
+    weight16_value = np.zeros((72, 5, 5, 72), dtype=np.int8)
     for i, j, k in np.ndindex((72, 5, 5)):
         weight16_value[i][j][k][i] = weight16_value_org[i][j][k][0]
     weight16.set_value(weight16_value)
@@ -260,7 +260,7 @@ def feature_extractor(act0, params,
     # [19] conv
     weight19 = ng.variable(dtype=weight_dtype, shape=(120, 5, 5, 120), name="layer3.0.1.layers.3.weight")
     weight19_value_org = params["layer3.0.1.layers.3.weight"]
-    weight19_value = np.zeros((120, 5, 5, 120))
+    weight19_value = np.zeros((120, 5, 5, 120), dtype=np.int8)
     for i, j, k in np.ndindex((120, 5, 5)):
         weight19_value[i][j][k][i] = weight19_value_org[i][j][k][0]
     weight19.set_value(weight19_value)
@@ -311,7 +311,7 @@ def feature_extractor(act0, params,
     # [23] conv
     weight23 = ng.variable(dtype=weight_dtype, shape=(120, 5, 5, 120), name="layer3.0.2.layers.3.weight")
     weight23_value_org = params["layer3.0.2.layers.3.weight"]
-    weight23_value = np.zeros((120, 5, 5, 120))
+    weight23_value = np.zeros((120, 5, 5, 120), dtype=np.int8)
     for i, j, k in np.ndindex((120, 5, 5)):
         weight23_value[i][j][k][i] = weight23_value_org[i][j][k][0]
     weight23.set_value(weight23_value)
@@ -363,7 +363,7 @@ def feature_extractor(act0, params,
     # [27] conv
     weight27 = ng.variable(dtype=weight_dtype, shape=(240, 5, 5, 240), name="layer4.0.0.layers.3.weight")
     weight27_value_org = params["layer4.0.0.layers.3.weight"]
-    weight27_value = np.zeros((240, 5, 5, 240))
+    weight27_value = np.zeros((240, 5, 5, 240), dtype=np.int8)
     for i, j, k in np.ndindex((240, 5, 5)):
         weight27_value[i][j][k][i] = weight27_value_org[i][j][k][0]
     weight27.set_value(weight27_value)
@@ -409,7 +409,7 @@ def feature_extractor(act0, params,
     # [30] conv
     weight30 = ng.variable(dtype=weight_dtype, shape=(480, 5, 5, 480), name="layer4.0.1.layers.3.weight")
     weight30_value_org = params["layer4.0.1.layers.3.weight"]
-    weight30_value = np.zeros((480, 5, 5, 480))
+    weight30_value = np.zeros((480, 5, 5, 480), dtype=np.int8)
     for i, j, k in np.ndindex((480, 5, 5)):
         weight30_value[i][j][k][i] = weight30_value_org[i][j][k][0]
     weight30.set_value(weight30_value)
@@ -461,7 +461,7 @@ def feature_extractor(act0, params,
     # [34] conv
     weight34 = ng.variable(dtype=weight_dtype, shape=(480, 5, 5, 480), name="layer4.0.2.layers.3.weight")
     weight34_value_org = params["layer4.0.2.layers.3.weight"]
-    weight34_value = np.zeros((480, 5, 5, 480))
+    weight34_value = np.zeros((480, 5, 5, 480), dtype=np.int8)
     for i, j, k in np.ndindex((480, 5, 5)):
         weight34_value[i][j][k][i] = weight34_value_org[i][j][k][0]
     weight34.set_value(weight34_value)
@@ -513,7 +513,7 @@ def feature_extractor(act0, params,
     # [38] conv
     weight38 = ng.variable(dtype=weight_dtype, shape=(480, 3, 3, 480), name="layer4.1.0.layers.3.weight")
     weight38_value_org = params["layer4.1.0.layers.3.weight"]
-    weight38_value = np.zeros((480, 3, 3, 480))
+    weight38_value = np.zeros((480, 3, 3, 480), dtype=np.int8)
     for i, j, k in np.ndindex((480, 3, 3)):
         weight38_value[i][j][k][i] = weight38_value_org[i][j][k][0]
     weight38.set_value(weight38_value)
@@ -559,7 +559,7 @@ def feature_extractor(act0, params,
     # [41] conv
     weight41 = ng.variable(dtype=weight_dtype, shape=(576, 3, 3, 576), name="layer4.1.1.layers.3.weight")
     weight41_value_org = params["layer4.1.1.layers.3.weight"]
-    weight41_value = np.zeros((576, 3, 3, 576))
+    weight41_value = np.zeros((576, 3, 3, 576), dtype=np.int8)
     for i, j, k in np.ndindex((576, 3, 3)):
         weight41_value[i][j][k][i] = weight41_value_org[i][j][k][0]
     weight41.set_value(weight41_value)
@@ -611,7 +611,7 @@ def feature_extractor(act0, params,
     # [45] conv
     weight45 = ng.variable(dtype=weight_dtype, shape=(576, 5, 5, 576), name="layer5.0.0.layers.3.weight")
     weight45_value_org = params["layer5.0.0.layers.3.weight"]
-    weight45_value = np.zeros((576, 5, 5, 576))
+    weight45_value = np.zeros((576, 5, 5, 576), dtype=np.int8)
     for i, j, k in np.ndindex((576, 5, 5)):
         weight45_value[i][j][k][i] = weight45_value_org[i][j][k][0]
     weight45.set_value(weight45_value)
@@ -657,7 +657,7 @@ def feature_extractor(act0, params,
     # [48] conv
     weight48 = ng.variable(dtype=weight_dtype, shape=(1152, 5, 5, 1152), name="layer5.0.1.layers.3.weight")
     weight48_value_org = params["layer5.0.1.layers.3.weight"]
-    weight48_value = np.zeros((1152, 5, 5, 1152))
+    weight48_value = np.zeros((1152, 5, 5, 1152), dtype=np.int8)
     for i, j, k in np.ndindex((1152, 5, 5)):
         weight48_value[i][j][k][i] = weight48_value_org[i][j][k][0]
     weight48.set_value(weight48_value)
@@ -709,7 +709,7 @@ def feature_extractor(act0, params,
     # [52] conv
     weight52 = ng.variable(dtype=weight_dtype, shape=(1152, 5, 5, 1152), name="layer5.0.2.layers.3.weight")
     weight52_value_org = params["layer5.0.2.layers.3.weight"]
-    weight52_value = np.zeros((1152, 5, 5, 1152))
+    weight52_value = np.zeros((1152, 5, 5, 1152), dtype=np.int8)
     for i, j, k in np.ndindex((1152, 5, 5)):
         weight52_value[i][j][k][i] = weight52_value_org[i][j][k][0]
     weight52.set_value(weight52_value)
@@ -761,7 +761,7 @@ def feature_extractor(act0, params,
     # [56] conv
     weight56 = ng.variable(dtype=weight_dtype, shape=(1152, 5, 5, 1152), name="layer5.0.3.layers.3.weight")
     weight56_value_org = params["layer5.0.3.layers.3.weight"]
-    weight56_value = np.zeros((1152, 5, 5, 1152))
+    weight56_value = np.zeros((1152, 5, 5, 1152), dtype=np.int8)
     for i, j, k in np.ndindex((1152, 5, 5)):
         weight56_value[i][j][k][i] = weight56_value_org[i][j][k][0]
     weight56.set_value(weight56_value)
@@ -813,7 +813,7 @@ def feature_extractor(act0, params,
     # [60] conv
     weight60 = ng.variable(dtype=weight_dtype, shape=(1152, 3, 3, 1152), name="layer5.1.0.layers.3.weight")
     weight60_value_org = params["layer5.1.0.layers.3.weight"]
-    weight60_value = np.zeros((1152, 3, 3, 1152))
+    weight60_value = np.zeros((1152, 3, 3, 1152), dtype=np.int8)
     for i, j, k in np.ndindex((1152, 3, 3)):
         weight60_value[i][j][k][i] = weight60_value_org[i][j][k][0]
     weight60.set_value(weight60_value)
